backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.95.2
uvicorn==0.23.2

# Database ORM and PostgreSQL driver
sqlalchemy==2.0.21
psycopg2-binary==2.9.9

# Authentication and hashing
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Configuration and validation
pydantic-settings==2.0.3
python-dotenv==1.0.0
pydantic==2.5.3
email-validator==2.1.1

# RabbitMQ integration
pika==1.3.2

# Redis integration
redis==4.6.0

# Optional: Gunicorn for production deployment
gunicorn==21.2.0
